feat: add buffer navigation and movement mappings
Added key mappings for navigating and moving buffer tabs. The new
mappings include "H" and "L", "<M-S-[>" and "<M-S-]>" for navigating to previous and next
buffers, and "<C-S-PageUp>", and "<C-S-PageDown>"
for moving buffer tabs left and right. This enhances buffer management
and improves user experience.
Unlike mini.bracketed buffer navigation mappings, these new ones are
based on the buffer position in the bufferline.